修复成交额值为(手)

+++ src/main/java/com/hao/datacollector/dto/quotation/HistoryTrendDTO.java
@@ -22,7 +22,7 @@ public class HistoryTrendDTO {
     private Double latestPrice;
 
     @Schema(description = "8.总成交量(手)1手=100股,总成交量不需要累加", example = "1423160")
-    private Long totalVolume;
+    private Double totalVolume;
 
     @Schema(description = "79.均价", example = "142.50")
     private Double averagePrice;
